Created scatterplot script
This script reads the output from the authorsFileTouches.py file. Using the csv file that is outputted from that file, this script plots the commits for each file in a project.

Each author is given a unique color. In the scatterplot, the x value denotes the file, and the y value denotes when in the project's lifecycle was a change committed.