Update functions and output
Much of the code was moved into functions. Added a "main" block towards the end of the file where the driving code is.

The scatterplot now has a legend with the color of the author, the author's name, and the amount of commits they made. The scatterplot is also saved as a file called Commits.png to better display the scatterplot and the legend.